+++ tus-clients/java/build.gradle
@@ -14,6 +14,10 @@ dependencies {
     implementation("io.github.cdimascio:dotenv-kotlin:6.4.1")
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+
+    // mockito dependencies
+    testImplementation 'org.mockito:mockito-core:3.12.4'
+    testImplementation 'org.mockito:mockito-junit-jupiter:3.12.4'
 }
 
 test {
